requirements/dashboard.txt
# Dashboard-specific dependencies
-r base.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
bcrypt>=4.0.0
apscheduler>=3.10.0
msgpack>=1.0.0
structlog>=23.0.0
